pyproject.toml
[project]
name = "multimodal-agent-builder"
version = "0.1.0"
description = "A multimodal AI agent builder supporting OpenAI GPT-4, Google Gemini, and Anthropic Claude"
authors = [{name = "J.M.Jarmacz", email = "webmaster@jarmacz.com"}]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    "openai>=1.0.0",
    "google-generativeai>=0.8.0",
    "anthropic>=0.64.0",
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "fastapi>=0.116.0",
    "uvicorn[standard]>=0.35.0",
    "pydantic>=2.11.0",
    "pydantic-settings>=2.10.0",
    "python-dotenv>=1.0.0",
    "Pillow>=11.0.0",
    "pydub>=0.25.1",
    "numpy>=2.0.0",
    "pandas>=2.0.0",
    "httpx>=0.28.0",
    "python-multipart>=0.0.5",
    "redis>=5.0.0",
    "pypdf>=4.0.0",
    "opentelemetry-sdk>=1.26.0",
    "opentelemetry-exporter-otlp>=1.26.0",
    "opentelemetry-instrumentation-fastapi>=0.47b0",
    "opentelemetry-instrumentation-httpx>=0.47b0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=1.0.0",
    "pytest-cov>=4.0.0",
    "black>=24.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "ipython>=8.0.0",
    "bandit[toml]>=1.7.9",
]

[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "config*"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312', 'py313']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "B", "C4", "D", "UP"]
ignore = ["D100", "D101", "D102", "D103", "D104", "D105", "D106", "D107"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "--verbose --cov=src --cov-report=term-missing"
asyncio_mode = "auto"
